Added new service provider Moved SpyOutput to output folder
