+++ src/Viserio/Component/Console/Provider/LazilyCommandsServiceProvider.php
@@ -1,0 +1,76 @@
+<?php
+declare(strict_types=1);
+namespace Viserio\Component\Console\Provider;
+
+use Interop\Container\ServiceProviderInterface;
+use Psr\Container\ContainerInterface;
+use Symfony\Component\Console\CommandLoader\ContainerCommandLoader;
+use Viserio\Component\Console\Application;
+use Viserio\Component\Contract\OptionsResolver\ProvidesDefaultOptions as ProvidesDefaultOptionsContract;
+use Viserio\Component\Contract\OptionsResolver\RequiresComponentConfig as RequiresComponentConfigContract;
+use Viserio\Component\OptionsResolver\Traits\OptionsResolverTrait;
+
+class LazilyCommandsServiceProvider implements
+    ServiceProviderInterface,
+    RequiresComponentConfigContract,
+    ProvidesDefaultOptionsContract
+{
+    use OptionsResolverTrait;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getFactories(): array
+    {
+        return [];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getExtensions(): array
+    {
+        return [
+            Application::class => [self::class, 'extendConsole'],
+        ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public static function getDimensions(): iterable
+    {
+        return ['viserio', 'console'];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public static function getDefaultOptions(): iterable
+    {
+        return [
+            'lazily_commands' => [],
+        ];
+    }
+
+    /**
+     * Extend viserio console with commands.
+     *
+     * @param \Psr\Container\ContainerInterface           $container
+     * @param null|\Viserio\Component\Console\Application $console
+     *
+     * @return null|\Viserio\Component\Console\Application
+     */
+    public static function extendConsole(
+        ContainerInterface $container,
+        ?Application $console = null
+    ): ?Application {
+        if ($console !== null) {
+            $options = self::resolveOptions($container);
+
+            $console->setCommandLoader(new ContainerCommandLoader($container, $options['lazily_commands']));
+        }
+
+        return $console;
+    }
+}
