+++ src/Viserio/Component/Console/Tests/Traits/ConfirmableTraitTest.php
@@ -6,7 +6,7 @@ use Narrowspark\TestingHelper\ArrayContainer;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Input\StringInput;
 use Viserio\Component\Console\Application;
-use Viserio\Component\Console\Tests\Fixture\SpyOutput;
+use Viserio\Component\Console\Output\SpyOutput;
 use Viserio\Component\Console\Tests\Fixture\ViserioConfirmableFalseCommand;
 use Viserio\Component\Console\Tests\Fixture\ViserioConfirmableTrueCommand;
