+++ src/Viserio/Component/Foundation/Console/Kernel.php
@@ -4,7 +4,6 @@ namespace Viserio\Component\Foundation\Console;
 
 use Closure;
 use Symfony\Component\Console\Command\Command as SymfonyCommand;
-use Symfony\Component\Console\CommandLoader\ContainerCommandLoader;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Debug\Exception\FatalThrowableError;
@@ -241,10 +240,6 @@ class Kernel extends AbstractKernel implements ConsoleKernelContract, Terminable
 
             $console->setVersion($this->resolvedOptions['version']);
             $console->setName($this->resolvedOptions['console_name']);
-
-            if ($container->has(ContainerCommandLoader::class)) {
-                $console->setCommandLoader($container->get(ContainerCommandLoader::class));
-            }
 
             foreach ($this->commands as $command) {
                 $console->add($container->resolve($command));
